+++ localflavor/au/forms.py
@@ -10,6 +10,8 @@ from django.forms.fields import CharField, RegexField, Select
 from django.utils.encoding import force_text
 from django.utils.translation import ugettext_lazy as _
 
+from localflavor.generic.forms import DeprecatedPhoneNumberFormFieldMixin
+
 from .au_states import STATE_CHOICES
 from .validators import AUBusinessNumberFieldValidator, AUTaxFileNumberFieldValidator
 
@@ -33,7 +35,7 @@ class AUPostCodeField(RegexField):
                                               max_length, min_length, *args, **kwargs)
 
 
-class AUPhoneNumberField(CharField):
+class AUPhoneNumberField(CharField, DeprecatedPhoneNumberFormFieldMixin):
     """
     A form field that validates input as an Australian phone number.