+++ localflavor/generic/deprecation.py
@@ -1,0 +1,2 @@
+class RemovedInLocalflavor20Warning(PendingDeprecationWarning):
+    pass