+++ localflavor/generic/forms.py
@@ -1,5 +1,10 @@
+from __future__ import unicode_literals
+
+import warnings
+
 from django import forms
 
+from localflavor.generic.deprecation import RemovedInLocalflavor20Warning
 from .validators import IBAN_COUNTRY_CODE_LENGTH, BICValidator, IBANValidator
 
 DEFAULT_DATE_INPUT_FORMATS = (
@@ -127,3 +132,14 @@ class BICFormField(forms.CharField):
         if value is not None:
             return value.upper()
         return value
+
+
+class DeprecatedPhoneNumberFormFieldMixin(object):
+    def __init__(self):
+        super(DeprecatedPhoneNumberFormFieldMixin, self).__init__()
+        warnings.warn(
+            "{} is deprecated in favor of the django-phonenumber-field library.".format(
+                self.__class__.__name__
+            ),
+            RemovedInLocalflavor20Warning,
+        )