+++ localflavor/id_/forms.py
@@ -11,6 +11,8 @@ from django.forms.fields import Field, Select
 from django.utils.encoding import force_text
 from django.utils.translation import ugettext_lazy as _
 
+from localflavor.generic.forms import DeprecatedPhoneNumberFormFieldMixin
+
 postcode_re = re.compile(r'^[1-9]\d{4}$')
 phone_re = re.compile(r'^(\+62|0)[2-9]\d{7,10}$')
 plate_re = re.compile(r'^(?P<prefix>[A-Z]{1,2}) ' +
@@ -57,7 +59,7 @@ class IDProvinceSelect(Select):
         super(IDProvinceSelect, self).__init__(attrs, choices=PROVINCE_CHOICES)
 
 
-class IDPhoneNumberField(Field):
+class IDPhoneNumberField(Field, DeprecatedPhoneNumberFormFieldMixin):
     """
     An Indonesian telephone number field.