+++ localflavor/il/forms.py
@@ -9,6 +9,7 @@ from django.forms.fields import Field, RegexField
 from django.utils.translation import ugettext_lazy as _
 
 from localflavor.generic.checksums import luhn
+from localflavor.generic.forms import DeprecatedPhoneNumberFormFieldMixin
 
 id_number_re = re.compile(r'^(?P<number>\d{1,8})-?(?P<check>\d)$')
 mobile_phone_number_re = re.compile(r'^(\()?0?(5[02-9])(?(1)\))-?\d{7}$')  # including palestinian mobile carriers
@@ -72,7 +73,7 @@ class ILIDNumberField(Field):
         return value
 
 
-class ILMobilePhoneNumberField(RegexField):
+class ILMobilePhoneNumberField(RegexField, DeprecatedPhoneNumberFormFieldMixin):
     """A form field that validates its input as an Israeli Mobile phone number."""
 
     default_error_messages = {