+++ localflavor/pk/models.py
@@ -1,6 +1,7 @@
 from django.db.models import CharField
 from django.utils.translation import ugettext_lazy as _
 
+from localflavor.generic.models import DeprecatedPhoneNumberField
 from . import forms
 from .pk_states import STATE_CHOICES
 
@@ -44,7 +45,7 @@ class PKPostCodeField(CharField):
         return super(PKPostCodeField, self).formfield(**defaults)
 
 
-class PKPhoneNumberField(CharField):
+class PKPhoneNumberField(CharField, DeprecatedPhoneNumberField):
     """A model field that checks that the value is a valid Pakistani phone number (nine to eleven digits)."""
 
     description = _("Pakistani Phone number")